heat

+++ soilFluxes3D/header/heat.h
@@ -7,5 +7,5 @@ double getNonIsothermalAirVaporConductivity(long i, double myTMean);
 void restoreHeat();
 void initializeBalanceHeat();
 void updateBalanceHeatWholePeriod();
-bool computeHeat(float myTime);
+bool computeHeat(double myTime);
